pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["D-Diyazzz <diyazzz.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
dotenv = "^0.0.5"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
greenlet = "^3.0.3"
pytz = "^2024.1"
python-multipart = "^0.0.9"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
anyio = "^4.3.0"
httpx = "^0.27.0"
redis = "^5.0.8"
fpdf = "^1.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
